EJB3/code/7_LarkU_EJB_WS/src/test/java/ttl/larku/service/CourseServiceTest.java
package ttl.larku.service;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

import ttl.larku.app.WeldUtil;
import ttl.larku.domain.Course;
import ttl.larku.domain.ScheduledClass;
import ttl.larku.domain.Student;

public class CourseServiceTest {

	private String name1 = "Bloke";
	private String name2 = "Blokess";
	private String newName = "Different Bloke";
	
	private String code1 = "BOT-101";
	private String code2 = "BOT-202";
	private String title1 = "Intro To Botany";
	private String title2 = "Outtro To Botany";
	
	private String startDate1 = "10/10/2012";
	private String startDate2 = "10/10/2013";
	private String endDate1 = "05/10/2013";
	private String endDate2 = "05/10/2014";
	
	private Course course1;
	private Course course2;
	private ScheduledClass class1;
	private ScheduledClass class2;
	private Student student1;
	private Student student2;
	
    private static WeldUtil wu = WeldUtil.INSTANCE;
	
	@BeforeClass
	public static void beforeClass() {
		wu = WeldUtil.INSTANCE;
	}

	private CourseService courseService;
	
	@Before
	public void setup() {
        courseService = wu.getBean (CourseService.class);
		student1 = new Student(name1);
		student2 = new Student(name2);
		course1 = new Course(code1, title1);
		course2 = new Course(code2, title2);
		
		class1 = new ScheduledClass(course1, startDate1, endDate1);
		class2 = new ScheduledClass(course2, startDate2, endDate2);
		
		courseService.clear();
	}
	
	@Test
	public void testCreateCourse() {
		Course newCourse = courseService.createCourse(code1, title1);
		
		Course result = courseService.getCourse(newCourse.getId());
		
		assertEquals(title1, result.getTitle());
		assertEquals(1, courseService.getAllCourses().size());
	}
	
	@Test
	public void testDeleteCourse() {
		Course course1 = courseService.createCourse(code1, title1);
		Course course2 = courseService.createCourse(code2, title2);
		
		assertEquals(2, courseService.getAllCourses().size());
		
		courseService.deleteCourse(course1.getId());
		
		assertEquals(1, courseService.getAllCourses().size());
		assertEquals(title2, courseService.getAllCourses().get(0).getTitle());
	}
	
	@Test
	public void testUpdateCourse() {
		Course course1 = courseService.createCourse(code1, title1);
		
		assertEquals(1, courseService.getAllCourses().size());
		
		course1.setTitle(title2);
		courseService.updateCourse(course1);
		
		assertEquals(1, courseService.getAllCourses().size());
		assertEquals(title2, courseService.getAllCourses().get(0).getTitle());
	}

}
